docker-4-5/votos/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
psycopg2-binary==2.9.7
redis==5.0.1
pydantic==2.5.0